Add files via upload
I worked on some skills and skill points. I added a lot more things like more getters and setters to the Player class and the Enemy class. We have some sort of simple AI for the enemy right now, although it's quite sad at the moment. I commented out anything that had to do with items since it wasn't compiling for me, so if you can get these classes to work and compile with working item classes, that would be amazing.